Refactor agent structure by integrating LangChain, enhancing Scratch programming assistance, and implementing error handling for JSON input; removed obsolete drag_agent and new_agent files, and streamlined block execution logic.
