+++ browseruse/Agent/utils/state.py
@@ -1,0 +1,3 @@
+class State(dict):
+    query: str
+    result: dict
