+++ browseruse/Agent/utils/tool.py
@@ -1,0 +1,54 @@
+import json
+from typing import List, Dict, Any
+from langchain_core.tools import tool
+import os
+import sys
+
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '../..')))
+
+
+from tools.execution import Executor
+
+executor = Executor()
+
+@tool
+def make_blocks(json_string: str) -> str:
+    """
+        Executes a sequence of block manipulation commands generated by an agent.
+
+        Arguments:
+            - json_plan (str): A JSON string containing a list of steps to execute.
+              Each step should specify:
+            - step (int): The step number in the sequence.
+            - category (str): The block category (e.g., "Events", "Control").
+            - block (str): The name of the block to interact with.
+
+        Example json_plan:
+        {
+          "steps": [
+            {
+              "step": 1,
+              "category": "Events",
+              "block": "when green flag clicked"
+            },
+            {
+              "step": 2,
+              "category": "Control",
+              "block": "forever"
+            }
+          ]
+        }
+
+        The function locates each block by fuzzy matching, clicks the category tab,
+        and performs drag-and-drop actions to build the desired program structure.
+        """
+    if len(json_string.strip()) == 0:
+        print("Empty JSON string provided.")
+        return "false"
+    try:
+        executor.executor_tool(json_string)
+        return "true"
+    except json.JSONDecodeError:
+        print("Invalid JSON format")
+        return "false"
